+++ History.md
@@ -4,6 +4,7 @@
 - Several styling improvements, especially for devices with small screens (e.g. smartphones)
 - The "My motions" and "My amendments" sections on the consultation overview now list all motions/amendments that one has created, ininiated or expressed (dis)like towards. Not only the ones ininiated as before. This makes it easier to check on motions/amendments that are for some reason not visible anymore on this site (withdrawn, merged, ...)
 - Bugfix: The proposed procedure overview for regular users was not accessible on devices with small screens
+- Long-broken references to OpenID-based login were removed
 
 ### Version 4.2.3 (2019-07-18)